# fast paths for static chunk writes + string/int fields